Factor out getPluralExamples()
