+++ frontend/src/core/plural/components/PluralSelector.js
@@ -53,37 +53,18 @@ export class PluralSelectorBase extends React.Component<InternalProps> {
     }
 
     render() {
-        const { pluralForm, locale } = this.props;
+        const props = this.props;
+        const { pluralForm } = props;
 
-        if (pluralForm === -1 || !locale || locale.cldrPlurals.length <= 1) {
+        if (pluralForm === -1 || !props.locale || props.locale.cldrPlurals.length <= 1) {
             return null;
         }
 
-        const nbPlurals = locale.cldrPlurals.length;
-        const examples = {};
-
-        if (nbPlurals === 2) {
-            examples[locale.cldrPlurals[0]] = 1;
-            examples[locale.cldrPlurals[1]] = 2;
-        }
-        else {
-            // This variable is used in the pluralRule we eval in the while block.
-            let n = 0;
-            while (Object.keys(examples).length < nbPlurals) {
-                // This `eval` is not so much evil. The pluralRule we parse
-                // comes from our database and is not a user input.
-                // eslint-disable-next-line
-                const rule = eval(locale.pluralRule);
-                if (!examples[locale.cldrPlurals[rule]]) {
-                    examples[locale.cldrPlurals[rule]] = n;
-                }
-                n++;
-            }
-        }
+        const examples = locale.getPluralExamples(props.locale);
 
         return <nav className="plural-selector">
             <ul>
-                { locale.cldrPlurals.map((item, i) => {
+                { props.locale.cldrPlurals.map((item, i) => {
                     return <li key={ item } className={ i === pluralForm ? 'active' : '' }>
                         <button onClick={ () => this.selectPluralForm(i) }>
                             <span>{ CLDR_PLURALS[item] }</span>
